Use docker to build the image
I would perfer to use podman, but podman causes a little problem. If
you're not root inside the container, you cannot easily write to
directories that are shared with the host system. There is a way to make
it happen, see https://www.tutorialworks.com/podman-rootless-volumes/.
But this solution mean that on the host you don't own the shared
directory anymore which makes writing to it on the host annoying. I also
don't want to be root inside the podman container, because of bitbakes
sanity checks (see last commit). Therefore the best option seems to be
to use docker, even though I perfer podman.